+++ catanatron_core/catanatron/__init__.py
@@ -3,6 +3,6 @@ This is to allow an API like:
 
 from catanatron import Game, Player, Color, Accumulator
 """
-from catanatron.game import Game, Accumulator
+from catanatron.game import Game, GameAccumulator
 from catanatron.models.player import Player, Color, RandomPlayer
 from catanatron.models.enums import Action, ActionType, WOOD, BRICK, SHEEP, WHEAT, ORE
