+++ tests/test_game.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 
 from catanatron.state_functions import get_actual_victory_points, player_has_rolled
 from catanatron.models.actions import maritime_trade_possibilities
-from catanatron.game import TURNS_LIMIT, Accumulator, Game
+from catanatron.game import Game
 from catanatron.state import (
     State,
     player_deck_replenish,
@@ -363,63 +363,6 @@ def test_second_placement_takes_cards_from_bank():
         game.play_tick()
 
     assert sum(game.state.resource_freqdeck) < 19 * 5
-
-
-def test_accumulators():
-    players = [
-        RandomPlayer(Color.RED),
-        RandomPlayer(Color.BLUE),
-        RandomPlayer(Color.WHITE),
-        RandomPlayer(Color.ORANGE),
-    ]
-    game = Game(players)
-
-    class MyAccumulator(Accumulator):
-        def __init__(self):
-            self.games = []
-            self.actions = []
-            self.initialized = False
-            self.finalized = False
-            self.final_game = None
-
-        def initialize(self, game):
-            self.initialized = True
-
-        def step(self, game, action):
-            self.games.append(game)
-            self.actions.append(action)
-
-        def finalize(self, game):
-            self.final_game = game
-            self.finalized = True
-
-    accumulator = MyAccumulator()
-    game.play(accumulators=[accumulator])
-
-    assert accumulator.initialized
-    assert len(accumulator.actions) == len(game.state.actions)
-    assert accumulator.finalized
-
-    # assert games in step() are before actions are taken
-    discard_actions = [
-        (i, a)
-        for i, a in enumerate(game.state.actions)
-        if a.action_type == ActionType.DISCARD
-    ]
-    for index, action in discard_actions:
-        game_snapshot = accumulator.games[index]
-        assert game_snapshot.state.is_discarding
-
-    # assert someone wins
-    assert accumulator.final_game is not None
-    points = [
-        get_actual_victory_points(accumulator.final_game.state, color)
-        for color in game.state.colors
-    ]
-    assert (
-        any([p >= 10 for p in points])
-        or accumulator.final_game.state.num_turns >= TURNS_LIMIT
-    )
 
 
 def test_vps_to_win_config():
