+++ tests/test_imports.py
@@ -1,10 +1,10 @@
-from catanatron import Game, RandomPlayer, Color, Accumulator
+from catanatron import Game, RandomPlayer, Color, GameAccumulator
 
 from catanatron_experimental.my_player import MyPlayer
 
 
 def test_top_level_imports_work():
-    class MyAccumulator(Accumulator):
+    class MyAccumulator(GameAccumulator):
         pass
 
     players = [
